move image cover update to handler from uploader
